requirements.txt
Django==1.6.7
Pillow==2.5.3
PyPDF2==1.23
South==1.0
argparse==1.2.1
django-dajax==0.9.2
django-dajaxice==0.6
django-excel-response==1.0
django-extensions==1.4.0
django-filter==0.7
django-selectable==0.8.0
django-selectable-select2==0.3.0
django-staticfiles-jquery==1.8.3
django-staticfiles-select2==3.4.5
django-webodt==0.3.1
html5lib==0.999
lxml==3.4.0
psycopg2==2.5.4
reportlab==3.1.8
six==1.8.0
wsgiref==0.1.2
xhtml2pdf==0.0.6